requirements.txt
ShopifyAPI
python-dotenv
requests
pymongo
boto3
fastapi
uvicorn
motor
awscli
